+++ internal/server/config/migrations_postgres/8_filter_presets.down.sql
@@ -1,0 +1,3 @@
+drop table if exists filter_presets;
+
+commit;